Fixed broken references to kernel normalizer headers, which have been moved to kernel/normalize
