+++ src/shogun/kernel/CommUlongStringKernel.cpp
@@ -10,7 +10,7 @@
 
 #include <shogun/lib/common.h>
 #include <shogun/kernel/CommUlongStringKernel.h>
-#include <shogun/kernel/SqrtDiagKernelNormalizer.h>
+#include <shogun/kernel/normalize/SqrtDiagKernelNormalizer.h>
 #include <shogun/features/StringFeatures.h>
 #include <shogun/io/SGIO.h>
