+++ src/shogun/kernel/FixedDegreeStringKernel.cpp
@@ -10,7 +10,7 @@
 
 #include <shogun/lib/common.h>
 #include <shogun/kernel/FixedDegreeStringKernel.h>
-#include <shogun/kernel/SqrtDiagKernelNormalizer.h>
+#include <shogun/kernel/normalize/SqrtDiagKernelNormalizer.h>
 #include <shogun/features/Features.h>
 #include <shogun/features/StringFeatures.h>
 #include <shogun/io/SGIO.h>
