+++ src/shogun/kernel/PolyKernel.cpp
@@ -13,7 +13,7 @@
 #include <shogun/lib/common.h>
 #include <shogun/io/SGIO.h>
 #include <shogun/kernel/PolyKernel.h>
-#include <shogun/kernel/SqrtDiagKernelNormalizer.h>
+#include <shogun/kernel/normalize/SqrtDiagKernelNormalizer.h>
 #include <shogun/features/DotFeatures.h>
 
 using namespace shogun;
